requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
pydantic==2.9.0
pinecone==6.0.0
groq==0.30.0
pypdf==3.17.1
